+++ LISEZ-MOI.md
@@ -4,11 +4,13 @@
 
 Le langage Python offre une opportunité sans précédent dans l'histoire de l'informatique: réunir et faire travailler sur les mêmes programmes des développeurs, des scientifiques et des étudiants de tout horizon. Parmi une longue liste de qualités cette particularité apparait comme l'une des raisons 'génétiques' à son succès qui ne fait que s'accroître et s'accélèrer à l'échelle mondiale.
 
-En sus de ses capacités dans les autres domaines comme internet et l'administration réseaux, Python joue aujourd'hui un rôle central en matière de traitement de données, de calcul, et d'inteligence artificielle. Au regard des intentions et des moyens mobilisés à son sujet ces dernières années par les grands noms de l'IT, nous assistons litéralement à la naissance d'une nouvelle ère pour Python, un nouvel empire déjà grandiose et remplis de promesses.
+En sus de ses capacités dans les autres domaines comme internet et l'administration des réseaux, Python joue aujourd'hui un rôle central en matière de traitement de données, de calcul, et d'inteligence artificielle. Au regard des intentions et des moyens mobilisés à son sujet par les grands noms de l'IT ces dernières années, nous assistons litéralement à la naissance d'une nouvelle ère pour Python, un nouvel empire déjà grandiose et remplis de promesses.
 
-Les industriels et autres curieux n'ont plus qu'à intégrer ce tumulte virvoltant se dressant au coeur d'une innovation flamboyante. Cette idée peut paraître saugrenue. Elle ne l'est pas. L'idée d'une intégration durable des compétences Python au sein des entreprises peut paraître irréaliste. Elle ne l'est pas. C'est même complétement souhaitable du point de vue de la productivité.
+Dans ce contexte les industriels et autres curieux en quête de productivité n'auraient donc plus qu'à intégrer ce tumulte virvoltant au coeur d'une innovation flamboyante? L'idée vous paraît saugrenue? Elle ne l'est pas. Elle est même à prendre très au sérieux.
+ 
+## low-code et minimalisme, l'alchimie gagnante
 
-## low-code et low-costs, les fondations d'un mariage durable
+En matière de développement informatique, deux tendances méritent une attention particulière: no-code et low-code. Ces terminologies commerciales éclairent sur certaines motivations profondent des grands éditeurs. Avec le principe no-code il nous est promis de tout concevoir sans une ligne de programmation. Le principe low-code quand à lui poursuit la même finalité en souplesse et en transparence avec un code concis, lisible, maintenable. Les évolutions en cours concernant les langages Python, Javascript et Rust, il est prévisible que ce dernier principe draine son lot de prouesses et d'inovations inatendues dans un avenir proche. Sous l'impulsion des GAFAM, le low-code est devenu une lame de fond ayant pour conséquence le renouveau à marche forcée des librairies logicielles open-source: faire mieux avec moins est une question de survie dans la compétition mondiale.
 
-En matière de développement informatique, deux tendances méritent une attention particulière: no-code et low-code. Ces terminologies commerciales éclairent sur certaines motivations profondent des grands éditeurs. Avec le principe no-code il nous est promis de tout pouvoir concevoir sans une ligne de programmation. Le principe low-code quand à lui poursuit la même finalité mais en souplesse et en transparence avec un code concis, lisible, maintenable. De part les évolutions actuelles des langages Python, Javascript et Rust, il est prévisible que ce dernier principe finisse par tenir rapidemment toute sa promesse. En réalité le low-code est déjà une lame de fond ayant pour conséquence le renouveau des librairies logicielles à marche forcée sous l'impulsion des GAFAM.
+Par nature l'informatique, et en particulier la programmation, nous apparaît souvent comme un sujet des plus sophistiqués, peu accessible, réservés aux experts et passionés. C'est oublier le bénéfice du bon sens en la matière. Un logiciel survendu nous apporte-il toujours plus qu'une macro bien sentie dans un tableur? Le temps passé à maitriser un logiciel spécifique doté de multiples fonctionalités nous est-il toujours si profitable? Ne rêvons-nous pas de la solution qui ferait juste ce qui nous rendrait service et de façon fiable?
