+++ datavision/data.csv
@@ -1,0 +1,5 @@
+FACEBOOK,30
+GITHUB,44
+GOOGLE,64
+TWITTER,17
+WEIBO,19
