+++ datavision/data.json
@@ -1,0 +1,7 @@
+{
+    "FACEBOOK": 30,
+    "GITHUB"  : 44,
+    "GOOGLE"  : 64,
+    "TWITTER" : 17,
+    "WEIBO"   : 19
+}
