+++ datavision/test.html
@@ -1,0 +1,76 @@
+<!DOCTYPE html>
+
+<h1>DataVision</h1>
+
+<!-- Load d3.js -->
+<script src="d3.min.js"></script>
+
+<!-- Create a div where the graph will take place -->
+<div id="account_group"></div>
+
+<script>
+
+let data = d3.json("data.json")
+
+let margin = {top: 20, right: 20, bottom: 30, left: 40};
+let svgWidth = 720, svgHeight = 300;
+let height = svgHeight- margin.top- margin.bottom, width = svgWidth - margin.left - margin.right;
+let sourceNames = [], sourceCount = [];
+
+let x = d3.scaleBand().rangeRound([0, width]).padding(0.1),
+    y = d3.scaleLinear().rangeRound([height, 0]);
+for(let key in data){
+    if(data.hasOwnProperty(key)){
+        sourceNames.push(key);
+        sourceCount.push(parseInt(data[key]));
+    }
+}
+x.domain(sourceNames);
+y.domain([0, d3.max(sourceCount, function(d) { return d; })]);
+
+let svg = d3.select("#account_group").append("svg");
+svg.attr('height', svgHeight)
+    .attr('width', svgWidth);
+
+svg = svg.append("g")
+         .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
+
+svg.append("g")
+    .attr("class", "axis axis--x")
+    .attr("transform", "translate(0," + height + ")")
+    .call(d3.axisBottom(x));
+
+svg.append("g")
+    .attr("class", "axis axis--y")
+    .call(d3.axisLeft(y).ticks(5))
+    ;
+        
+// Create rectangles
+let bars = svg.selectAll('.bar')
+    .data(sourceNames)
+    .enter()
+    .append("g");
+
+bars.append('rect')
+    .attr('class', 'bar')
+    .attr("x", function(d) { return x(d); })
+    .attr("y", function(d) { return y(data[d]); })
+    .attr("width", x.bandwidth())
+    .attr("height", function(d) { return height - y(data[d]); });
+    
+bars.append("text")
+    .text(function(d) { 
+        return data[d];
+    })
+    .attr("x", function(d){
+        return x(d) + x.bandwidth()/2;
+    })
+    .attr("y", function(d){
+        return y(data[d]) - 5;
+    })
+    .attr("font-family" , "sans-serif")
+    .attr("font-size" , "14px")
+    .attr("fill" , "black")
+    .attr("text-anchor", "middle");
+
+</script>
