+++ sweetheart/globals.py
@@ -1,6 +1,53 @@
 
-import os,sys
+import os,sys,subprocess
 from collections import UserDict
+
+
+class sp:
+    """ namespace providing basic subprocess features """
+
+    run = lambda *args,**kwargs: subprocess.run(args,**kwargs)
+    shell = lambda str,**kwargs: subprocess.run(str,**kwargs,shell=True)
+
+    @classmethod
+    def terminal(cls,cmd:str,select:str):
+        """ run cmd within selected terminal 
+            select must be in xterm|winterm|wsl """
+
+        wsl = f"cmd.exe /c start wsl {cmd} &"
+        winterm = f"cmd.exe /c start wt wsl {cmd} &"
+        xterm = f"xterm -C -geometry 190x19 -e {cmd} &"
+
+        assert select in "xterm|winterm|wsl"
+        cls.shell(eval(select))
+        
+    @classmethod
+    def poetry(cls,*args,**kwargs):
+        return cls.run(BaseConfig.poetry_bin,*args,**kwargs)
+
+    @classmethod
+    def python(cls,*args,**kwargs):
+        return cls.run(BaseConfig.python_bin,*args,**kwargs)
+
+    @classmethod
+    def set_python_env(cls,path:str):
+        """ get python venv path from poetry and set it within config
+            beware that path must exist and contain a poetry project """
+
+        os.chdir(path)
+        venv = cls.poetry("env","info","--path",
+            text=True,capture_output=True).stdout.strip()
+
+        if venv == "":
+            raise Exception("Error, no python env found")
+
+        echo("set python env:",venv)
+        BaseConfig.python_bin = f"{venv}/bin/python"
+
+    @classmethod
+    def exit(cls):
+        """ stop and exit in good way """
+        sys.exit()
 
 
 # set default configuration
@@ -11,14 +58,14 @@ class BaseConfig(UserDict):
     label = "sweetheart"
 
     # get environment settings
-    USER = os.environ['USER']
-    HOME = os.environ['HOME']
     PWD = os.environ['PWD']
+    HOME = os.environ['HOME']
+    USER = os.environ['USER'].capitalize()
     WSL_DISTRO_NAME = os.getenv('WSL_DISTRO_NAME')
 
     # default paths settings
     poetry_bin = f"{HOME}/.poetry/bin/poetry"
-    python_bin = "python3"# venv disabled
+    python_bin = "sws python"# unknown venv
 
     def __init__(self,project) -> None:
 
@@ -30,7 +77,6 @@ class BaseConfig(UserDict):
         # default sandbox settings
         self.is_webapp = True
         self.is_mongodb_local = True
-        self.is_uvicorn_local = True
         self.is_jupyter_local = False
         self.is_cherrypy_local = False
 
@@ -57,12 +103,12 @@ class BaseConfig(UserDict):
 
             "templates_dir": "templates",
             "templates_settings": {
-                "_default_libs_": "knacss py",
+                "_default_libs_": "py",
                 "_async_": self.async_host,
                 "_static_": "",# ""=disabled
             },
             "static_files": {
-                "/favicon.ico": "resources/favicon.ico",
+                #"/favicon.ico": "resources/favicon.ico",
             },
             "static_dirs": {
                 "/resources": "resources",
@@ -72,20 +118,19 @@ class BaseConfig(UserDict):
             },
         }
 
-        # set a default html welcome message
-        self.welcome = f"""
-          <div style="text-align:center;">
-            <h1><br><br>Welcome {self.USER.capitalize()} !<br><br></h1>
-            <h3>sweetheart</h3>
-            <p>a supercharged heart for the non-expert hands</p>
-            <p>that will give you full power at the speedlight</p>
-            <p><a href="documentation/index.html"
-                class="btn" role="button">Get Started Now!</a></p>
-            <p><br>or code immediately using <a href="jupyter">JupyterLab</a></p>
-            <p><br><br><em>this message appears because there
-                was nothing else to render here</em></p>
-          </div>
-            """
+
+# set a default html welcome message
+WELCOME = f"""
+  <div style="text-align:center;font-size:1.2em;">
+    <h1><br><br>Welcome {BaseConfig.USER} !<br><br></h1>
+    <h3>sweetheart</h3>
+    <p>a supercharged heart for the non-expert hands</p>
+    <p>that will give you full power at the speedlight</p>
+    <p><a href="documentation/index.html">Get Started Now!</a></p>
+    <p><br>or code immediately using <a href="jupyter">JupyterLab</a></p>
+    <p><br><br><em>this message appears because there
+      was nothing else to render here</em></p>
+  </div>"""
 
 
 # provide convenient functions for givin messages
